pyproject.toml
[project]
name = "fastapi-testovoe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "greenlet>=3.3.0",
    "pydantic-settings>=2.12.0",
    "pytest-asyncio>=1.3.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
